rename components dir
